+++ js/data.js
@@ -1,0 +1,58 @@
+export {
+  OFFER_TITLES,
+  OFFER_SECRIPTIONS,
+  OFFER_TYPES,
+  OFFER_CHECK_IN_OUT,
+  OFFER_FEATURES,
+  OFFER_PHOTOS
+};
+
+const OFFER_TITLES = [
+  'Объявление №1',
+  'Объявление №2',
+  'Объявление №3',
+  'Объявление №4',
+  'Объявление №5',
+  'Объявление №6',
+  'Объявление №7',
+  'Объявление №8',
+  'Объявление №9',
+  'Объявление №10',
+];
+
+const OFFER_SECRIPTIONS = [
+  'Евроремонт',
+  'Прекрасный вид из окна',
+  'Хорошее соотношение цена / качество',
+  'Историческая часть города',
+  'Метро рядом',
+];
+
+const OFFER_TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const OFFER_CHECK_IN_OUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const OFFER_FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+];
+
+const OFFER_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+
